feat(router): refactor
 - On sort la partie Road de la partie Router
 - Création du package server/road
 - Chaque section possède son fichier dans /road/
 - Les méthodes sont exportées
